4A done
need to further tweak params, also make rat faster

+++ Rat/Core/Inc/controller.h
@@ -8,6 +8,6 @@
 #include "main.h"
 
 void move(int8_t n);
-void turn(int8_t n);
+void turn(int n);
 
 #endif /* INC_CONTROLLER_H_ */
